Move get batch size into utils and added some tests
